+++ test/escape.test.ts
@@ -14,3 +14,24 @@ Deno.test("Escape filter", async () => {
     expected: "&lt;h1&gt;Hello world&lt;/h1&gt;",
   });
 });
+
+Deno.test("Escape by default", async () => {
+  await test({
+    options: {
+      autoescape: true,
+    },
+    template: `
+    {{ "<h1>Hello world</h1>" }}
+    `,
+    expected: "&lt;h1&gt;Hello world&lt;/h1&gt;",
+  });
+  testSync({
+    options: {
+      autoescape: true,
+    },
+    template: `
+    {{ "<h1>Hello world</h1>" }}
+    `,
+    expected: "&lt;h1&gt;Hello world&lt;/h1&gt;",
+  });
+});