+++ CHANGELOG.md
@@ -4,6 +4,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [0.8.0] - Unreleased
+### Added
+- New `echo` tag [#14].
+- New `unescape`filter.
+- New option `autoescape`.
+
+### Removed
+- The `raw` tag. Use `echo` instead.
+
 ## [0.7.3] - 2023-09-02
 ### Fixed
 - Escape backslashes [#12], [#13].
@@ -88,7 +97,9 @@ First version
 [#10]: https://github.com/oscarotero/vento/issues/10
 [#12]: https://github.com/oscarotero/vento/issues/12
 [#13]: https://github.com/oscarotero/vento/issues/13
+[#14]: https://github.com/oscarotero/vento/issues/14
 
+[0.8.0]: https://github.com/oscarotero/vento/compare/v0.7.3...HEAD
 [0.7.3]: https://github.com/oscarotero/vento/compare/v0.7.2...v0.7.3
 [0.7.2]: https://github.com/oscarotero/vento/compare/v0.7.1...v0.7.2
 [0.7.1]: https://github.com/oscarotero/vento/compare/v0.7.0...v0.7.1